Upgrade and Refactor Project for comprehension
