+++ Project/src/ImGui/ImGuiContext.cpp
@@ -1,0 +1,171 @@
+//
+// Created by ianpo on 18/07/2024.
+//
+
+#include "ImGui/ImGuiContext.hpp"
+
+
+#include "glad/glad.h"
+
+#define GLFW_INCLUDE_NONE 1
+#define IMGUI_IMPL_OPENGL_LOADER_CUSTOM 1
+
+#include <imgui.h>
+#include <imgui_impl_opengl3.h>
+#include <imgui_impl_glfw.h>
+
+#include <GLFW/glfw3.h>
+
+#include "Core/Logger.hpp"
+
+namespace PC {
+	ImGuiContext::ImGuiContext(std::weak_ptr<Window> window) : m_Window(std::move(window)) {
+		auto wd = m_Window.lock();
+		if (!wd) {
+			PC_ERROR("Window do not exist. Cannot create the ImGui context.");
+			return;
+		}
+		static const char *const glsl_version = "#version 460";
+		// Setup Dear ImGui context
+		IMGUI_CHECKVERSION();
+		ImGui::CreateContext();
+		ImGuiIO &io = ImGui::GetIO();
+		(void) io;
+		io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;// Enable Keyboard Controls
+		io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad; // Enable Gamepad Controls
+		io.ConfigFlags |= ImGuiConfigFlags_DockingEnable;    // Enable Docking
+		io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;  // Enable Multi-Viewport / Platform Windows
+		//io.ConfigViewportsNoAutoMerge = true;
+		//io.ConfigViewportsNoTaskBarIcon = true;
+
+		// Setup Dear ImGui style
+		ImGui::StyleColorsDark();
+		//ImGui::StyleColorsLight();
+
+		// When viewports are enabled we tweak WindowRounding/WindowBg so platform windows can look identical to regular ones.
+		ImGuiStyle &style = ImGui::GetStyle();
+		if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable) {
+			style.WindowRounding = 0.0f;
+			style.Colors[ImGuiCol_WindowBg].w = 1.0f;
+		}
+
+		// Setup Platform/Renderer backends
+		ImGui_ImplGlfw_InitForOpenGL(wd->GetNative<GLFWwindow>(), true);
+		ImGui_ImplOpenGL3_Init(glsl_version);
+	}
+
+	ImGuiContext::~ImGuiContext() {
+	}
+	void ImGuiContext::UseDockspace() {
+		// READ THIS !!!
+		// TL;DR; this demo is more complicated than what most users you would normally use.
+		// If we remove all options we are showcasing, this demo would become:
+		//     void ShowExampleAppDockSpace()
+		//     {
+		//         ImGui::DockSpaceOverViewport(ImGui::GetMainViewport());
+		//     }
+		// In most cases you should be able to just call DockSpaceOverViewport() and ignore all the code below!
+		// In this specific demo, we are not using DockSpaceOverViewport() because:
+		// - (1) we allow the host window to be floating/moveable instead of filling the viewport (when opt_fullscreen == false)
+		// - (2) we allow the host window to have padding (when opt_padding == true)
+		// - (3) we expose many flags and need a way to have them visible.
+		// - (4) we have a local menu bar in the host window (vs. you could use BeginMainMenuBar() + DockSpaceOverViewport()
+		//      in your code, but we don't here because we allow the window to be floating)
+
+		const static bool opt_fullscreen = true;
+		const static bool opt_padding = false;
+
+		static ImGuiDockNodeFlags dockspace_flags = ImGuiDockNodeFlags_None;
+
+		// We are using the ImGuiWindowFlags_NoDocking flag to make the parent window not dockable into,
+		// because it would be confusing to have two docking targets within each others.
+		ImGuiWindowFlags window_flags = ImGuiWindowFlags_MenuBar | ImGuiWindowFlags_NoDocking;
+		if (opt_fullscreen) {
+			const ImGuiViewport *viewport = ImGui::GetMainViewport();
+			ImGui::SetNextWindowPos(viewport->WorkPos);
+			ImGui::SetNextWindowSize(viewport->WorkSize);
+			ImGui::SetNextWindowViewport(viewport->ID);
+			ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 0.0f);
+			ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize, 0.0f);
+			window_flags |= ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove;
+			window_flags |= ImGuiWindowFlags_NoBringToFrontOnFocus | ImGuiWindowFlags_NoNavFocus;
+		} else {
+			dockspace_flags &= ~ImGuiDockNodeFlags_PassthruCentralNode;
+		}
+
+		// When using ImGuiDockNodeFlags_PassthruCentralNode, DockSpace() will render our background
+		// and handle the pass-thru hole, so we ask Begin() to not render a background.
+		if (dockspace_flags & ImGuiDockNodeFlags_PassthruCentralNode)
+			window_flags |= ImGuiWindowFlags_NoBackground;
+
+		// Important: note that we proceed even if Begin() returns false (aka window is collapsed).
+		// This is because we want to keep our DockSpace() active. If a DockSpace() is inactive,
+		// all active windows docked into it will lose their parent and become undocked.
+		// We cannot preserve the docking relationship between an active window and an inactive docking, otherwise
+		// any change of dockspace/settings would lead to windows being stuck in limbo and never being visible.
+		if (!opt_padding)
+			ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0.0f, 0.0f));
+		ImGui::Begin("DockSpace", nullptr, window_flags);
+		if (!opt_padding)
+			ImGui::PopStyleVar();
+
+		if (opt_fullscreen)
+			ImGui::PopStyleVar(2);
+
+		// Submit the DockSpace
+		ImGuiIO &io = ImGui::GetIO();
+		if (io.ConfigFlags & ImGuiConfigFlags_DockingEnable) {
+			ImGuiID dockspace_id = ImGui::GetID("MyDockSpace");
+			ImGui::DockSpace(dockspace_id, ImVec2(0.0f, 0.0f), dockspace_flags);
+		}
+
+
+		if (ImGui::BeginMenuBar()) {
+			if (ImGui::BeginMenu("Options")) {
+				// Disabling fullscreen would allow the window to be moved to the front of other windows,
+				// which we can't undo at the moment without finer window depth/z control.
+				if (auto window = m_Window.lock()) {
+					if (ImGui::MenuItem("Exit")) {
+						window->Close();
+					}
+				}
+				//			ImGui::MenuItem("Padding", NULL, &opt_padding);
+				ImGui::Separator();
+
+				ImGui::EndMenu();
+			}
+
+			ImGui::EndMenuBar();
+		}
+
+		ImGui::End();
+	}
+
+	void ImGuiContext::BeginFrame() {
+		// Start the Dear ImGui frame
+		ImGui_ImplOpenGL3_NewFrame();
+		ImGui_ImplGlfw_NewFrame();
+		ImGui::NewFrame();
+	}
+	void ImGuiContext::EndFrame() {
+		// Rendering
+		ImGui::Render();
+
+		ImGuiIO &io = ImGui::GetIO();
+		if (auto window = m_Window.lock()) {
+			io.DisplaySize = ImVec2(window->GetWidth(), window->GetHeight());
+		}
+
+		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
+
+		// Update and Render additional Platform Windows
+		// (Platform functions may change the current OpenGL context, so we save/restore it to make it easier to paste this code elsewhere.
+		//  For this specific demo app we could also call glfwMakeContextCurrent(window) directly)
+		if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable) {
+			GLFWwindow *backup_current_context = glfwGetCurrentContext();
+			ImGui::UpdatePlatformWindows();
+			ImGui::RenderPlatformWindowsDefault();
+			glfwMakeContextCurrent(backup_current_context);
+		}
+	}
+}
